+++ src/district_energy_model/dem_techs.py
@@ -16,6 +16,7 @@ from district_energy_model import dem_constants as C
 #------------------------------------------------------------------------------
 # Generation
 from district_energy_model.techs.dem_tech_grid_supply import GridSupply
+from district_energy_model.techs.dem_tech_grid_export import GridExport
   
 #------------------------------------------------------------------------------
 # Conversion